refactor: update storage instance attach info request size
Sets the request size value to uint64 as per PR feedback. This will
handle all possible values that can be seen.

+++ domain/application/service/storage/attachsize.go
@@ -13,7 +13,7 @@ import (
 // provisioned sizes are used when set; otherwise the requested size is used.
 // If the kind is unknown, the requested size is used.
 func CalculateStorageInstanceSizeForAttachment(info internal.StorageInstanceInfo) uint64 {
-	fallback := uint64(max(info.RequestedSizeMIB, 0))
+	fallback := max(info.RequestedSizeMIB, 0)
 
 	switch info.Kind {
 	case domainstorage.StorageKindFilesystem:
